Last saved state
Keeping track of last used directory for dialog boxes and at the same time reset them to a default state after changing projects has been a challenge. The problem is due to the separation of the project load handling and the io_manager itself. This unfortunately is needed as there is even more editors in the works that will rely on this. It's just a mind bender, that's all.

Took 1 hour 0 minutes